Added recyler view implementation

+++ app/src/main/java/com/example/sleepez/SleepData.java
@@ -51,7 +51,7 @@ public class SleepData {
         this.wakeTime = wakeTime;
     }
 
-    public double getSleepQuality() {
+    public int getSleepQuality() {
         return sleepQuality;
     }
